+++ polygraph/types/api.py
@@ -1,0 +1,2 @@
+def typedef(type_):
+    return type_.__type
