+++ polygraph/types/tests/test_object_type.py
@@ -1,10 +1,10 @@
 from unittest import TestCase
 
 from polygraph.exceptions import PolygraphValueError
-from polygraph.types.basic_type import NonNull
 from polygraph.types.decorators import field
 from polygraph.types.object_type import ObjectType
 from polygraph.types.scalar import Int, String
+from polygraph.types.type_builder import NonNull
 
 
 class HelloWorldObject(ObjectType):
